requirements.txt

Flask==1.1.2
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
pipenv==2020.6.2
pipreqs==0.4.10
pytesseract==0.3.4
requests==2.23.0
virtualenv==20.0.23
virtualenv-clone==0.5.4
gunicorn==20.0.4
